legacy_pipelines/vectorize/requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
requests>=2.28.0
numpy<2
